Remove TimerBroadcast
This was added briefly because the timer was
pausing when the app was in the background (sometimes).
I read somewhere using a BroadcastReceiver prevents the timer
from being slept by Android, which turned out to be false (on a Pixel
    4).
The actual solution was disabling battery optimizations, so this
broadcast receiver is now redundant.