Add one rep max calculator for best graphs
I tried out a bunch of formulas as well as
having them as options, and I ended up liking
the Brzycki formula the most.
https://en.wikipedia.org/wiki/One-repetition_maximum#Brzycki
All of them produced similar trends (for me) and the one using
exponents wouldn't work with the SQLite version on android
(can't use POWER function). Also having all the options looked
kind of cluttered. If people ask for it I'll add the other ones
later.